transfer course detail activity to fragment
